Alluxiofs performance benchmark (#34)
Refactor alluxiofs performance benchmark
