+++ benchmark/traffic_pattern/ray_pytorch_jpeg.csv
