+++ benchmark/traffic_pattern/ray_pytorch_jpeg_4workers.csv
